Backend/requirements.txt
fastapi
uvicorn
ollama
pypdf
faiss-cpu
python-multipart
pydantic
